requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord.py @ git+https://github.com/Rapptz/discord.py@00d39704da538969fc07fa4db71c18b2c62141b3
docutils==0.15.2
et-xmlfile==1.0.1
futures==3.1.1
google==3.0.0
goslate==1.5.1
idna==2.8
jdcal==1.4.1
matplotlib==3.1.2
mock==2.0.0
multidict==4.7.6
numpy==1.18.1
oauthlib==3.1.0
openpyxl==3.0.2
pandas==0.25.3
pbr==5.4.5
Pillow==7.2.0
pycipher==0.5.2
PyDictionary==2.0.1
pygame==1.9.6
Pygments==2.4.2
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.13.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
soupsieve==2.0.1
tweepy==3.9.0
typing-extensions==3.7.4.2
urbandict==0.6.1
urbandictionary==1.1
urllib3==1.25.6
virtualenv==16.7.7
Vocabulary==1.0.4
websockets==8.1
xlwt==1.3.0
yarl==1.5.1
mysql-connector-python==8.0.21
